MB-10800: avoid overflow computing cb_cond_timedwait wakeup time
tv_sec is usually long. Which is usually 32-bit on 32-bit
machines. And multiplying it by 1000, even if final result is
uint64_t, is still performed as 32-bit computation on 32-bit boxes.

Clearly, multiplying current unix-epoch time by 1000 overflows 32-bit.

The fix is to simply convert tv_sec to uint64_t before multiplication
which ensures that it's computed using 64-bit arithmetics.

Reviewed-on: http://review.couchbase.org/35919
Tested-by: buildbot <build@couchbase.com>

+++ src/cb_pthreads.c
@@ -143,7 +143,7 @@ void cb_cond_timedwait(cb_cond_t *cond, cb_mutex_t *mutex, unsigned int ms)
      * to convert back to an absolute time...
      */
     gettimeofday(&tp, NULL);
-    wakeup = (tp.tv_sec * 1000) + (tp.tv_usec / 1000) + ms;
+    wakeup = ((uint64_t)(tp.tv_sec) * 1000) + (tp.tv_usec / 1000) + ms;
     /* Round up for sub ms */
     if ((tp.tv_usec % 1000) > 499) {
         ++wakeup;
